Moved admin code verification out of LocalAdminApi
